# BikeWorldServer/api-docs.yaml
openapi: 3.0.0
info:
  title: bike-world API
  version: '2.0'
servers:
  - url: https://bike-world-ids.herokuapp.com/
paths:
...
  /api/v1/users/login:
    post:
      summary: Authenticate and get a new token.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: 'user1'
                password:
                  type: string
                  example: 'user_password'
      responses:
        '200':
          description: Token sucessfully created
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: true
                  message: 
                    type: string
                    example: 'Token sucessfully created'
                  token: 
                    type: string
                    example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJ0b3B0YWwuY29tIiwiZXhwIjoxNDI2NDIwODAwLCJodHRwOi8vdG9wdGFsLmNvbS9qd3RfY2xhaW1zL2lzX2FkbWluIjp0cnVlLCJjb21wYW55IjoiVG9wdGFsIiwiYXdlc29tZSI6dHJ1ZX0.yRQYnWzskCZUxPwaQupWkiUzKELZ49eM7oWxAQK_ZXw'
                  permissions: 
                    type: boolean
                    example: false
                  username: 
                    type: string
                    example: 'user1'
                  id: 
                    type: string
                    example: '627fb15a85bebd50ab3197a5'
                  self: 
                    type: string
                    example: '/api/v1/users/627fb15a85bebd50ab3197a5'
        '400':
          description: User not found.
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string  
                    example: 'Authentication failed. User not found.'
        '400':
          description: Bad Request. Check docs for required parameters. /api/v1/api-docs
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'Bad Request. Check docs for required parameters. /api/v1/api-docs'
        '400':
          description: Wrong username or password.
          content: 
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'Authentication failed. Wrong username or password.' 
  /api/v1/users/signUp:
    post:
      summary: Creates a new bike-world accont.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: 'user1'
                email:
                  type: string
                  example: 'user1@domain.com'
                password:
                  type: string
                  example: 'user_password'
                target:
                  type: string
                  example: 'principiante'
      responses:
        '201':
          description: Account successfully created.
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: true
                  message: 
                    type: string
                    example: 'Signup completed!'
                  token: 
                    type: string
                    example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJ0b3B0YWwuY29tIiwiZXhwIjoxNDI2NDIwODAwLCJodHRwOi8vdG9wdGFsLmNvbS9qd3RfY2xhaW1zL2lzX2FkbWluIjp0cnVlLCJjb21wYW55IjoiVG9wdGFsIiwiYXdlc29tZSI6dHJ1ZX0.yRQYnWzskCZUxPwaQupWkiUzKELZ49eM7oWxAQK_ZXw'
                  permissions: 
                    type: boolean
                    example: false
                  username: 
                    type: string
                    example: 'user1'
                  id: 
                    type: string
                    example: '627fb15a85bebd50ab3197a5'
                  self: 
                    type: string
                    example: '/api/v1/users/627fb15a85bebd50ab3197a5'
        '400':
          description: Bad Request. Check docs for required parameters. /api/v1/api-docs
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'Bad Request. Check docs for required parameters. /api/v1/api-docs'
        '409':
          description: Signup failed. User already exists.
          content: 
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'Signup failed. User already exists.' 
...
  /api/v1/users/{id}:
    put:
      summary: Update User info.
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the user to update.
          schema:
            type: string
            example: '627fb15a85bebd50ab3197a5'
        - name: x-access-token
          in: header
          required: true
          description: The jwt of the session.
          schema:
            type: string
            example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJ0b3B0YWwuY29tIiwiZXhwIjoxNDI2NDIwODAwLCJodHRwOi8vdG9wdGFsLmNvbS9qd3RfY2xhaW1zL2lzX2FkbWluIjp0cnVlLCJjb21wYW55IjoiVG9wdGFsIiwiYXdlc29tZSI6dHJ1ZX0.yRQYnWzskCZUxPwaQupWkiUzKELZ49eM7oWxAQK_ZXw'
      responses:
        '200':
          description: Get User info.
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: true
                  message: 
                    type: string
                    example: 'Information updated!'
                  token: 
                    type: string
                    description: The new token, since the user changed.
                    example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJ0b3B0YWwuY29tIiwiZXhwIjoxNDI2NDIwODAwLCJodHRwOi8vdG9wdGFsLmNvbS9qd3RfY2xhaW1zL2lzX2FkbWluIjp0cnVlLCJjb21wYW55IjoiVG9wdGFsIiwiYXdlc29tZSI6dHJ1ZX0.yRQYnWzskCZUxPwaQupWkiUzKELZ49eM7oWxAQK_ZXw'
                  permissions: 
                    type: boolean
                    example: false
                  username: 
                    type: string
                    example: 'user1'
                  id: 
                    type: string
                    example: '627fb15a85bebd50ab3197a5'
                  self: 
                    type: string
                    example: '/api/v1/users/627fb15a85bebd50ab3197a5'
        '400':
          description: Bad Request. Check docs for required parameters. /api/v1/api-docs
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'Bad Request. Check docs for required parameters. /api/v1/api-docs'
        '403':
          description: Unauthorized. You can update only your informations.
          content: 
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'Unauthorized. You can access only your informations.' 
        '401':
          description: Unauthorized. No token provided.
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'No token provided.'
        '403':
          description: Forbidden. Failed to authenticate token.
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'Failed to authenticate token.'                    

  ###################
  # route bikes
  ###################
...
  /api/v2/users/login:
    post:
      summary: Authenticate and get a new token.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: 'user1'
                password:
                  type: string
                  example: 'user_password'
      responses:
        '200':
          description: Token sucessfully created
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: true
                  message: 
                    type: string
                    example: 'Token sucessfully created'
                  token: 
                    type: string
                    example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJ0b3B0YWwuY29tIiwiZXhwIjoxNDI2NDIwODAwLCJodHRwOi8vdG9wdGFsLmNvbS9qd3RfY2xhaW1zL2lzX2FkbWluIjp0cnVlLCJjb21wYW55IjoiVG9wdGFsIiwiYXdlc29tZSI6dHJ1ZX0.yRQYnWzskCZUxPwaQupWkiUzKELZ49eM7oWxAQK_ZXw'
                  permissions: 
                    type: boolean
                    example: false
                  username: 
                    type: string
                    example: 'user1'
                  id: 
                    type: string
                    example: '627fb15a85bebd50ab3197a5'
                  self: 
                    type: string
                    example: '/api/v2/users/627fb15a85bebd50ab3197a5'
        '400':
          description: User not found
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string  
                    example: 'Authentication failed. Wrong username or password.'
        '400':
          description: Bad Request. Check docs for required parameters. /api/api-docs
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'Bad Request. Check docs for required parameters. /api/api-docs'
        '400':
          description: Wrong username or password.
          content: 
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'Authentication failed. Wrong username or password.' 
  /api/v2/users/signUp:
    post:
      summary: Creates a new bike-world accont.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: 'user1'
                email:
                  type: string
                  example: 'user1@domain.com'
                password:
                  type: string
                  example: 'user_password'
                target:
                  type: string
                  example: 'principiante'
      responses:
        '201':
          description: Account successfully created.
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: true
                  message: 
                    type: string
                    example: 'Signup completed!'
                  token: 
                    type: string
                    example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJ0b3B0YWwuY29tIiwiZXhwIjoxNDI2NDIwODAwLCJodHRwOi8vdG9wdGFsLmNvbS9qd3RfY2xhaW1zL2lzX2FkbWluIjp0cnVlLCJjb21wYW55IjoiVG9wdGFsIiwiYXdlc29tZSI6dHJ1ZX0.yRQYnWzskCZUxPwaQupWkiUzKELZ49eM7oWxAQK_ZXw'
                  permissions: 
                    type: boolean
                    example: false
                  username: 
                    type: string
                    example: 'user1'
                  id: 
                    type: string
                    example: '627fb15a85bebd50ab3197a5'
                  self: 
                    type: string
                    example: '/api/v2/users/627fb15a85bebd50ab3197a5'
        '400':
          description: Bad Request. Check docs for required parameters. /api/api-docs
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'Bad Request. Check docs for required parameters. /api/api-docs'
        '409':
          description: Signup failed. User already exists.
          content: 
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'Signup failed. User already exists.' 
...
  /api/v2/users/{id}:
    put:
      summary: Update User info.
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the user to update.
          schema:
            type: string
            example: '627fb15a85bebd50ab3197a5'
        - name: x-access-token
          in: header
          required: true
          description: The jwt of the session.
          schema:
            type: string
            example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJ0b3B0YWwuY29tIiwiZXhwIjoxNDI2NDIwODAwLCJodHRwOi8vdG9wdGFsLmNvbS9qd3RfY2xhaW1zL2lzX2FkbWluIjp0cnVlLCJjb21wYW55IjoiVG9wdGFsIiwiYXdlc29tZSI6dHJ1ZX0.yRQYnWzskCZUxPwaQupWkiUzKELZ49eM7oWxAQK_ZXw'
      responses:
        '200':
          description: Get User info.
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: true
                  message: 
                    type: string
                    example: 'Information updated!'
                  token: 
                    type: string
                    description: The new token, since the user changed.
                    example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJ0b3B0YWwuY29tIiwiZXhwIjoxNDI2NDIwODAwLCJodHRwOi8vdG9wdGFsLmNvbS9qd3RfY2xhaW1zL2lzX2FkbWluIjp0cnVlLCJjb21wYW55IjoiVG9wdGFsIiwiYXdlc29tZSI6dHJ1ZX0.yRQYnWzskCZUxPwaQupWkiUzKELZ49eM7oWxAQK_ZXw'
                  permissions: 
                    type: boolean
                    example: false
                  username: 
                    type: string
                    example: 'user1'
                  id: 
                    type: string
                    example: '627fb15a85bebd50ab3197a5'
                  self: 
                    type: string
                    example: '/api/v2/users/627fb15a85bebd50ab3197a5'
        '400':
          description: Bad Request. Check docs for required parameters. /api/api-docs
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'Bad Request. Check docs for required parameters. /api/api-docs'
        '403':
          description: Unauthorized. You can update only your informations.
          content: 
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'Unauthorized. You can access only your informations.' 
        '401':
          description: Unauthorized. No token provided.
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'No token provided.'
        '403':
          description: Forbidden. Failed to authenticate token.
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'Failed to authenticate token.'                    

  ###################
  # route bikes
  ###################
...